requirements.txt
Flask==3.0.0
Flask-CORS==6.0.0
gunicorn==23.0.0
boto3==1.34.0
python-nmap==0.7.1